+++ src/aux/point.rs
@@ -2,11 +2,11 @@ use std::clone::*;
 
 #[derive(Copy, Clone)]
 pub struct Point {
-    x: f64,
-    y: f64,
+    x: i32,
+    y: i32,
 }
 
-pub fn new_res_point(x: f64, y: f64) -> Box<Point> {
+pub fn new_res_point(x: i32, y: i32) -> Box<Point> {
     let p = Point { x, y };
 
     Box::new(p)
@@ -14,8 +14,8 @@ pub fn new_res_point(x: f64, y: f64) -> Box<Point> {
 
 pub fn new_res_point_from_angle(angle: f64, dist: f64) -> Box<Point> {
     let theta: f64 = 2.0 * std::f64::consts::PI * angle;
-    let x: f64 = dist * theta.cos();
-    let y: f64 = dist * theta.sin();
+    let x: i32 = (dist * theta.cos()) as i32;
+    let y: i32 = (dist * theta.sin()) as i32;
 
     new_res_point(x, y)
 }
